+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### `jsonschema-generator`
+#### Changed
+- BREAKING CHANGE: minimum Java version `17`
+- BREAKING CHANGE: minimum Jackson version `3.x`
+
 ### `jsonschema-module-jackson`
 #### Fixed
 - support `@JacksonAnnotationsInside` annotated combo annotations also when looking for `@JsonPropertyDescription`